feat: add scheduler evaluation
