+++ scripts/MathAgent_task.txt
@@ -1,0 +1,30 @@
+Mark has 4 bags of marbles, each with 25 marbles. He gives 3 marbles to each of his 5 friends. How many marbles does he have left?
+Sara bakes 60 cookies and divides them evenly among her 4 friends. How many cookies does each friend receive?
+A farmer has 5 fields with 120 apple trees each. If each tree produces 30 apples, how many apples does the farmer have in total?
+Jenny reads 150 pages of a book each day. If she continues at this rate, how many pages will she read in 10 days?
+Alex buys 3 packs of pens, with each pack containing 12 pens, and gives 5 pens away. How many pens does Alex have now?
+A school orders 200 notebooks, each containing 50 pages. If students take 3 pages each day from their notebooks, how many days will the pages last?
+A baker makes 10 cakes, each cut into 8 slices. If he sells 45 slices, how many full cakes does he have left?
+There are 6 boxes with 18 chocolates each. If 4 chocolates are eaten from each box, how many chocolates are left in total?
+A library has 8 shelves, each with 35 books. If 15 books are checked out from each shelf, how many books remain in the library?
+Paul has 150 stamps and decides to distribute them equally among his 7 friends. How many stamps does each friend receive, and how many are left over?
+A train travels 450 miles in 5 hours. If it maintains the same speed, how far will it travel in 8 hours?
+A class of 30 students needs 5 pencils each for an exam. If each box contains 50 pencils, how many boxes are needed?
+A gardener plants 200 seeds. If 85% of the seeds germinate, how many plants does he expect to grow?
+An office has 9 computers, each running 4 applications simultaneously. How many applications are running in total?
+Emma saves $75 each month. After how many months will she have saved enough to buy a bike for $450?
+A marathon runner completes 3/5 of a race in 2 hours. At the same pace, how long will it take to finish the race?
+A box contains 250 screws. If one piece of furniture requires 15 screws, how many pieces of furniture can be assembled with the screws in the box?
+A movie theater has 15 rows with 20 seats each. If 70% of the seats are occupied, how many people are in the theater?
+If a car travels at 60 miles per hour and increases its speed by 5 miles per hour each hour, what is its speed after 4 hours?
+A wall needs 120 bricks to be built completely. If a worker can install 15 bricks per hour, how many hours will it take to build the wall?
+A recipe calls for 250 grams of flour, which is 25% of the total ingredients' weight. What is the total weight of all ingredients?
+During a sale, a shirt’s price is reduced by 15%, making the new price $85. What was the original price?
+A parking lot has 100 spaces and is 75% full. How many spaces are currently occupied, and how many are available?
+A water tank holds 500 liters and is currently 80% full. How many more liters can it hold?
+A runner increases his distance by 10% each day, starting with 5 km on day one. How far will he run on the fifth day?
+A cube's volume is 125 cubic inches. What is the length of each side of the cube?
+A factory produces 400 units of a product each day. If each unit sells for $25, what is the total daily revenue?
+A bus travels 300 km with 10 stops along the way. If each stop lasts 5 minutes, how long is the total stop time?
+A book has 20 chapters, each with 15 pages. How many pages does the book have?
+A store has 300 apples and sells 85% of them. How many apples are sold, and how many are left?
