+++ scripts/NarrativeAgent_task.txt
@@ -1,0 +1,1 @@
+Craft a tale about a valiant warrior on a quest to uncover priceless treasures hidden within a mystical island.
