+++ scripts/RecAgent_task.txt
@@ -1,0 +1,1 @@
+I want to take a tour to New York during the spring break, recommend some restaurants around for me.
