+++ src/llm_kernel/llm_config/gemma-1.1-2b-it.json
@@ -1,5 +1,5 @@
 {
     "model_type": "causal_lm",
     "open_sourced": true,
-    "model_name": "google/gemma-2b-it"
+    "model_name": "google/gemma-1.1-2b-it"
 }
